requirements.txt
mysql-connector-python==8.0.28
python-decouple==3.6
